% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateby.notified.cases.R
\name{aggregateby.notified.cases}
\alias{aggregateby.notified.cases}
\title{Aggregate total notified cases per epiweekyear}
\usage{
aggregateby.notified.cases(df.in, current.epiweek, current.epiyear)
}
\arguments{
\item{df.in}{Data frame object with first column providing location id and second column providing epiweekyear}

\item{current.epiweek}{Integer value specifing epiweek last entry to be considered. E.g, if latest epiyearweek is 2014W32,
current.epiweek should be 32.}

\item{current.epiyear}{Integer value specifing epiyear last entry to be considered. E.g, if latest epiyearweek is 2014W32,
current.epiweek should be 2014.}
}
\value{
Data frame with aggregate counts by epiyearweek
}
\description{
Aggregate total notified cases per epiweekyear
}

