% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay.profile.R
\name{delay.profile}
\alias{delay.profile}
\title{Delay Profile}
\usage{
delay.profile(dt, q = c(0.05, 0.25, 0.5, 0.75, 0.95))
}
\arguments{
\item{dt}{Notification table with at least two columns identifying locality[[1]] and delay[[2]]}

\item{q}{List of quantiles to calculate. Default c(.5, .25, .5, .75, .95)}
}
\value{
named list with quantiles by locality
}
\description{
Delay Profile
}
