% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay.profile.timeseries.R
\name{delay.profile.timeseries}
\alias{delay.profile.timeseries}
\title{Title delay.profile.timeseries}
\usage{
delay.profile.timeseries(dt, trainning = NULL)
}
\arguments{
\item{dt}{data frame with at least the following ordered columns: location id [,1], epiweek [,2],
and delay [,3] columns}

\item{trainning}{How many weeks from start of time series to use as training for
baseline quantiles for dealy distribution. Default: 52.}
}
\value{
Function \code{delay.profile.timeseries} returns a data.frame with quantiles
for historical delay by windows of 4, 8, 26, and 52 weeks.
Colum baseweek identify week of distribution calculation.
Columns start and end indicates notification date interval.
So, for baseweek 2014W01, start 2013W43, and end 2013W47 would report the
delay distribution for cases notified between 2013W43 and 2013W47 to be used as reference
for week 2014W01 delay pattern with a window of 4 weeks.
}
\description{
Method to evaluate how delay profile changes through time, with varying time windows.
}
