% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episem.R
\name{episem}
\alias{episem}
\title{Function written by Oswaldo C.
hosted at http://www.procc.fiocruz.br/~oswaldo/ocmisc.r}
\usage{
episem(x, separa = "W", retorna = "YW")
}
\arguments{
\item{x}{Date to be converted to epidemiological week. Can be of typy Date or character in the format 'YYYY-MM-DD'}

\item{separa}{Character to separate YYYY and WW.}

\item{retorna}{What should be return, if epidemiological year and week ('YW'), epi. year only ('Y') or epi. week only ('W').
Default: 'YW'.}
}
\description{
Function written by Oswaldo C.
hosted at http://www.procc.fiocruz.br/~oswaldo/ocmisc.r
}
\author{
Oswaldo G. Cruz
}
