% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.columns.from.date.R
\name{generate.columns.from.date}
\alias{generate.columns.from.date}
\title{Generate new columns epiyearweek, epiweek and epiyear from date column}
\usage{
generate.columns.from.date(x)
}
\arguments{
\item{x}{String with date in the format YYYY-DD-MM or an object of class Date}
}
\value{
\code{generate.columns.from.date} returns a named list with the values
epiweek, epiyear and epiyearweek
epiyearweek have the format 2009W42
}
\description{
Function \code{generate.columns.from.date} creates new column from original one
written in the format YYYY-MM-DD, using Brazilian epidemiological week system
}
\examples{
df <- data.frame(list(date=c('2009-08-01', '2009-09-01', '2009-10-01', '2010-01-01')))
t(sapply(df$date, generate.columns.from.date))

}

