% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.columns.from.epiyearweek.R
\name{generate.columns.from.epiyearweek}
\alias{generate.columns.from.epiyearweek}
\title{Generate new columns epiweek and epiyear from epiyearweek}
\usage{
generate.columns.from.epiyearweek(x)
}
\arguments{
\item{x}{String with date in the format YYYY-DD-MM or an object of class Date}
}
\value{
\code{generate.columns.from.date} returns a named list with the values
epiweek, epiyear
}
\description{
Function \code{generate.columns.from.epiyearweek} creates new column from original one
written in the format YYYY*WW
}
\examples{
df <- data.frame(list(epiyearweek=c('2009W30', '2009W35', '2009W39', '2009W52')))
t(sapply(df$epiyearweek, generate.columns.from.epiyearweek))

}

