% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastepiweek.R
\name{lastepiweek}
\alias{lastepiweek}
\title{Calculate number of year's last epidemiological week using Brazilian standard.}
\usage{
lastepiweek(ano)
}
\arguments{
\item{ano}{Year}
}
\description{
Calculate number of year's last epidemiological week using Brazilian standard.
}
\keyword{export}

