% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opportunity.example.data-data.R
\docType{data}
\name{opportunity.example.data}
\alias{opportunity.example.data}
\title{Example data of case registry with notification and digitalization dates}
\format{An object of class \code{"data.frame"}.}
\usage{
data(opportunity.example.data)
}
\description{
Data from SARI cases registered in the Brazilian database for a
specific state. Includes cases notified between 2013-2014, and
digitized up to last epidemiological week of 2014.
}
\examples{
data(opportunity.example.data)
table(opportunity.example.data)
}
\keyword{datasets}
