% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous.epiyearweek.from.date.R
\name{previous.epiyearweek.from.date}
\alias{previous.epiyearweek.from.date}
\title{Returns previous Brazilian epiweek from a given day.}
\usage{
previous.epiyearweek.from.date(day)
}
\arguments{
\item{day}{String with date in the format YYYY-DD-MM or an object of class Date}
}
\value{
string with epiweek
}
\description{
Function \code{previous.epiyearweek.from.date} calculates the previous Brazilian
epiweek from a given date passed in the format YYYY-MM-DD
}
\examples{
day <- '2010-03-27'
previous.epiyearweek.from.date(day)

}
